Bug fixes for VR/BR sorting. Restructured some components.
