+++ src/components/common/player-item/index.ts
@@ -1,0 +1,3 @@
+import PlayerItem from "./PlayerItem";
+
+export default PlayerItem;
