--- src/components/common/search/index.ts
@@ -1,3 +1,0 @@
-import Search from "./Search";
-
-export default Search;
