+++ src/components/players/Players.tsx
@@ -1,9 +1,9 @@
 import { Paper, PaperProps, Stack } from "@mantine/core";
 
+import PlayerItem from "../common/player-item/PlayerItem";
 import React from "react";
 import { Room } from "@/lib/types";
-import RoomPlayer from "../rooms/RoomPlayer";
-import usePlayersFilter from "@/lib/hooks/usePlayersFilter";
+import usePlayersFilter from "@/lib/hooks/filters/usePlayersFilter";
 
 interface Props extends PaperProps {
 	rooms: Room[];
@@ -21,7 +21,7 @@ const Players = ({ rooms, ...props }: Props) => {
 		>
 			<Stack gap={0}>
 				{data.map((player, index) => (
-					<RoomPlayer
+					<PlayerItem
 						key={player.fc}
 						player={player}
 						filled={index % 2 === 1}
