--- src/components/players/form/sortByData.ts
@@ -1,16 +1,0 @@
-const sortByData: { label: string; value: string }[] = [
-	{
-		label: "Name",
-		value: "name",
-	},
-	{
-		label: "VR",
-		value: "vr",
-	},
-	{
-		label: "BR",
-		value: "br",
-	},
-];
-
-export default sortByData;
