+++ src/components/rooms/form/index.ts
@@ -1,6 +1,3 @@
 import SortForm from "./SortForm";
-import sortByData from "./sortByData";
-
-export { sortByData as sortPropertyData };
 
 export default SortForm;
