--- src/components/rooms/form/sortByData.ts
@@ -1,24 +1,0 @@
-const sortByData: { label: string; value: string }[] = [
-	{
-		label: "Room name",
-		value: "name",
-	},
-	{
-		label: "Lifetime",
-		value: "lifetime",
-	},
-	{
-		label: "Player count",
-		value: "players",
-	},
-	{
-		label: "Average VR",
-		value: "vr",
-	},
-	{
-		label: "Average BR",
-		value: "br",
-	},
-];
-
-export default sortByData;
